backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pymongo==4.8.0
pydantic==2.8.2
requests==2.32.3
beautifulsoup4==4.12.3
google-generativeai==0.7.2
python-dotenv==1.0.1